Feat/homeiscrizione

+++ database/migrations/2026_04_29_080933_add_batch_uuid_column_to_activity_log_table.php
@@ -8,15 +8,25 @@ class AddBatchUuidColumnToActivityLogTable extends Migration
 {
     public function up()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->uuid('batch_uuid')->nullable()->after('properties');
-        });
+        $connection = config('activitylog.database_connection');
+        $table = config('activitylog.table_name');
+
+        if (! Schema::connection($connection)->hasColumn($table, 'batch_uuid')) {
+            Schema::connection($connection)->table($table, function (Blueprint $table) {
+                $table->uuid('batch_uuid')->nullable()->after('properties');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->dropColumn('batch_uuid');
-        });
+        $connection = config('activitylog.database_connection');
+        $table = config('activitylog.table_name');
+
+        if (Schema::connection($connection)->hasColumn($table, 'batch_uuid')) {
+            Schema::connection($connection)->table($table, function (Blueprint $table) {
+                $table->dropColumn('batch_uuid');
+            });
+        }
     }
 }
